+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 
-- ongoing
+# v0.16.4 - Adding measurements
+  - Expose mac-addresses for network and zigbee devices
+  - Expose min/max thermostat (and heater) values and resolution (step in HA)
   - Changed mac-addresses in userdata/fixtures to be obfuscated but unique 
+
+# v0.16.3 - Typing
+  - Code quality improvements
 
 # v0.16.2 - Generic and Stretch
   - As per Core deprecation of python 3.8, removed CI/CD testing and bumped pypi to 3.9 and production